chore: fix seeds by expecting tenant on Tenants.run_migrations?
Seeds were failing with the following error:

    ** (KeyError) key :migrations_ran not found in: nil

    If you are using the dot syntax, such as map.field, make sure the left-hand side of the dot is a map
        (realtime 2.34.53) lib/realtime/tenants.ex:304: Realtime.Tenants.run_migrations?/1
        (realtime 2.34.53) lib/realtime/tenants/migrations.ex:163: Realtime.Tenants.Migrations.run_migrations/1
        (ecto_sql 3.11.3) lib/ecto/adapters/sql.ex:1358: anonymous fn/3 in Ecto.Adapters.SQL.checkout_or_transaction/4
        (db_connection 2.7.0) lib/db_connection.ex:1756: DBConnection.run_transaction/4
        priv/repo/seeds.exs:13: (file)

The issue was that running Cache.get_tenant_by_external_id would be called by Cachex
in a separate process which would use a different Database connection
than the seed Repo.transaction. Making so that a Tenant that just got
created during the seed process could not be found.